create components folder
