--- app/ProductList/ProductList.module.css
@@ -1,14 +1,0 @@
-.list {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-around;
-    border: 1px solid #ccc;
-}
-
-.list > ul > li {
-    list-style: none;
-    padding: 10px;
-    margin: 0 24px;
-    border-top: 1px solid #ccc;
-}
